third_party/WebKit/LayoutTests/fast/viewport/viewport-75-expected.txt

CONSOLE MESSAGE: line 3: The key "x;" is not recognized and ignored.
CONSOLE MESSAGE: line 3: Error parsing a meta element's content: ';' is not a valid key-value pair separator. Please use ',' instead.
ALERT: viewport size 64x70.4 scale 5 with limits [5, 5] and userScalable true
